Refactor transactional mutation actions and move builder helper classes into a separate directory
