+++ lib/dynamoid/transactions/mutation/delete_with_primary_key.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative 'base'
-require_relative 'delete_request_builder'
+require_relative 'builders/delete_request_builder'
 
 module Dynamoid
   module Transactions
@@ -36,7 +36,7 @@ module Dynamoid
         end
 
         def action_requests
-          builder = DeleteRequestBuilder.new(@model_class)
+          builder = Builders::DeleteRequestBuilder.new(@model_class)
           builder.hash_key = cast_and_dump(@model_class.hash_key, @hash_key)
           builder.range_key = cast_and_dump(@model_class.range_key, @range_key) if @model_class.range_key?
