+++ lib/dynamoid/transactions/mutation/save.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative 'base'
-require_relative 'update_request_builder'
+require_relative 'builders/update_request_builder'
 
 module Dynamoid
   module Transactions
@@ -139,7 +139,7 @@ module Dynamoid
           changes = @model.attributes.slice(*@model.changed.map(&:to_sym))
           changes_dumped = Dynamoid::Dumping.dump_attributes(changes, @model_class.attributes)
 
-          builder = UpdateRequestBuilder.new(@model_class)
+          builder = Builders::UpdateRequestBuilder.new(@model_class)
           builder.hash_key = dump_attribute(@model_class.hash_key, @model.hash_key)
           builder.range_key = dump_attribute(@model_class.range_key, @model.range_value) if @model_class.range_key?
